requirements.txt
# Flask核心依赖
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0

# 数据库依赖
PyMySQL==1.1.0
SQLAlchemy==2.0.23

# 安全和验证
marshmallow==3.20.1
Flask-Limiter==3.5.0

# 工具库
python-dotenv==1.0.0
Werkzeug==2.3.7

# 生产环境依赖
gunicorn==21.2.0
redis==5.0.1

# 监控和日志
sentry-sdk[flask]==1.38.0
psutil==5.9.6
